separated out header and nav components for more specificity in terms of what each component file does
