+++ frontend/src/Components/00-Header/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const HeaderCont = styled.header`
+  display: flex;
+  justify-content: space-between;
+  /* padding: 2em 2em 2em 2em; */
+  background-color: lightgray;
+
+  .app-name {
+    font-size: 2.25rem;
+  }
+`;
