+++ frontend/src/Components/01-Navbar/Navbar.js
@@ -1,19 +1,11 @@
 import React from 'react';
-import { AppName, LogInBtn, NavbarCont, SignUpBtn } from './style';
+import { NavbarCont } from './style';
 
 export default function Navbar() {
   return (
     <NavbarCont>
-      <div className="left">
-        <AppName>Sleep Tracker</AppName>
-        <div>Features</div>
-      </div>
-      <div className="right">
-        {/* <div>Log In</div>
-        <div>Sign Up</div> */}
-        <LogInBtn>Log In</LogInBtn>
-        <SignUpBtn>Sign Up</SignUpBtn>
-      </div>
+      <div>About</div>
+      <div>Features</div>
     </NavbarCont>
   );
 }
