+++ frontend/src/Components/02-SignUpLogInBtns/SignUpLoginBtns.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { SignUpBtnCont, LogInBtnCont } from './style';
+
+export const SignUpBtn = () => {
+  return <SignUpBtnCont>Sign Up</SignUpBtnCont>;
+};
+
+export const LogInBtn = () => {
+  return <LogInBtnCont>Log In</LogInBtnCont>;
+};
