+++ frontend/src/Components/02-SignUpLogInBtns/style.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+const Button = styled.button`
+  background: none;
+  font-weight: bold;
+  border: none;
+  font-size: 1.125rem;
+  font-family: inherit;
+`;
+
+export const LogInBtnCont = styled(Button)``;
+
+export const SignUpBtnCont = styled(Button)`
+  border: thin black solid;
+  border-radius: 0.875rem;
+  padding: 0.5rem 0.875rem 0.5rem 0.875rem;
+`;
