+++ frontend/src/Components/04-LandingPage/LandingPage.js
@@ -1,17 +1,18 @@
 import React from 'react';
 import { MainContent } from './style';
-import Navbar from '../01-Navbar/Navbar';
-import Footer from '../02-Footer/Footer';
+import Header from '../00-Header/Header';
+// import Navbar from '../01-Navbar/Navbar';
+import Footer from '../03-Footer/Footer';
 
 export default function LandingPage() {
   return (
     <>
-      <Navbar />
-      <MainContent>
+      <Header />
+      {/* <MainContent>
         <div>Call to action slogan here.</div>
         <button>Get Started</button>
       </MainContent>
-      <Footer />
+      <Footer /> */}
     </>
   );
 }
